B1: Changed to double to avoid implicit int to double conversion

+++ src/benchmarks/b1/b1_2_for_each_quadratic.h
@@ -8,11 +8,15 @@
 
 
 namespace B1 {
+
+    //TODO: double
+
+
     template<class OuterExecutionPolicy, class InnerExecutionPolicy>
     inline void
     b1_2_for_each_quadratic(OuterExecutionPolicy &&outerExecutionPolicy,
                             InnerExecutionPolicy &&innerExecutionPolicy,
-                            const std::vector<int> &input_data) {
+                            const std::vector<double> &input_data) {
 
         // quite straight forward std::for_each. The body of the lambda has no intention behind it, the goal was just
         // to have a big enough computation that takes some amount of time.
